Change db model

+++ api/wams/services/int11_planner.py
@@ -44,31 +44,39 @@ def insert_into_planner(dict):
 
 def get_planner(from_date, to_date):
 
-    Planner.objects.all().delete()
     
     payload = {
         "from_date": from_date,
         "to_date": to_date
     }
 
+    # print("eeeeeeee")
     r = requests.post("http://174.138.28.157/getPlanner.php", data=payload)
+    # print("qqqqqqqq")
+    print(r)
 
     json_dictionary = json.loads(r.content)
-    for key in json_dictionary:
-        if (key == "results"):
-            print(key, ":", json_dictionary[key])
-            if (type(json_dictionary[key]) == dict):
-                # return single json
-                print("dict")
-                insert_into_planner(json_dictionary[key])
-            elif (type(json_dictionary[key]) == list):
-                # return array of json
-                print("list")
-                results_json = json_dictionary[key]
-                for x in results_json:
-                    insert_into_planner(x)
+    print(json_dictionary)
 
-    return json.loads(r.content)
+    if json_dictionary : 
+
+        Planner.objects.all().delete()
+        
+        for key in json_dictionary:
+            if (key == "results"):
+                print(key, ":", json_dictionary[key])
+                if (type(json_dictionary[key]) == dict):
+                    # return single json
+                    print("dict")
+                    insert_into_planner(json_dictionary[key])
+                elif (type(json_dictionary[key]) == list):
+                    # return array of json
+                    print("list")
+                    results_json = json_dictionary[key]
+                    for x in results_json:
+                        insert_into_planner(x)
+
+        return json.loads(r.content)
 
     # wsdl = "https://pasb-dev-uwa-iws.oracleindustry.com/ouaf/webservices/CM-PLANNER?WSDL"
     # session = Session()
